chore: update ic-agent to `0.17.0`

This is the minimal set of changes necessary to update ic-agent to `0.17.0`

update `k256` and `p256` as a consequence (cargo doesn't like multiple versions of these crates)

update some cycles to `u128` as needed by the newer version of `ic-agent` 

See merge request dfinity-lab/public/ic!5411

+++ rs/tests/src/util.rs
@@ -91,7 +91,7 @@ impl<'a> UniversalCanister<'a> {
     pub async fn new_with_comp_alloc(
         agent: &'a Agent,
         compute_allocation: Option<u64>,
-        cycles: Option<u64>,
+        cycles: Option<u128>,
     ) -> Result<UniversalCanister<'a>, String> {
         let payload = universal_canister_argument_builder().stable_grow(1).build();
 
@@ -116,7 +116,7 @@ impl<'a> UniversalCanister<'a> {
         Ok(Self { agent, canister_id })
     }
 
-    pub async fn new_with_cycles<C: Into<u64>>(
+    pub async fn new_with_cycles<C: Into<u128>>(
         agent: &'a Agent,
         cycles: C,
     ) -> UniversalCanister<'a> {
